Conflicts:
	src/main/java/org/olat/repository/_i18n/LocalStrings_de.properties
	src/main/java/org/olat/repository/controllers/RepositoryMainController.java

+++ src/main/java/org/olat/catalog/CatalogManager.java
@@ -26,6 +26,7 @@
 package org.olat.catalog;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.Iterator;
 import java.util.List;
 
@@ -90,9 +91,8 @@ public class CatalogManager extends BasicManager implements UserDataDeletable, I
 	 * [spring]
 	 * @param userDeletionManager
 	 */
-	private CatalogManager(UserDeletionManager userDeletionManager) {
+	private CatalogManager() {
 		// singleton
-		userDeletionManager.registerDeletableUserData(this);
 		catalogManager = this;
 	}
 
@@ -109,6 +109,20 @@ public class CatalogManager extends BasicManager implements UserDataDeletable, I
 	public CatalogEntry createCatalogEntry() {
 		return new CatalogEntryImpl();
 	}
+	
+	public List<CatalogEntry> getChildNodesOf(CatalogEntry ce) {
+		StringBuilder sb = new StringBuilder();
+		sb.append("select cei from ").append(CatalogEntryImpl.class.getName()).append(" as cei ")
+		  .append(" inner join fetch cei.ownerGroup as ownerGroup")
+		  .append(" inner join fetch cei.parent as parentCei")
+		  .append(" inner join fetch parentCei.ownerGroup as parentOwnerGroup")
+		  .append(" where parentCei.key=:parentKey");
+
+		return dbInstance.getCurrentEntityManager()
+				.createQuery(sb.toString(), CatalogEntry.class)
+				.setParameter("parentKey", ce.getKey())
+				.getResultList();
+	}
 
 	/**
 	 * Children of this CatalogEntry as a list of CatalogEntries
@@ -119,40 +133,31 @@ public class CatalogManager extends BasicManager implements UserDataDeletable, I
 		return getChildrenOf(ce, 0, -1, CatalogEntry.OrderBy.name, true);
 	}
 
+	/**
+	 * Return the nodes
+	 * @param ce
+	 * @param firstResult
+	 * @param maxResults
+	 * @param orderBy
+	 * @param asc
+	 * @return
+	 */
 	public List<CatalogEntry> getChildrenOf(CatalogEntry ce, int firstResult, int maxResults, CatalogEntry.OrderBy orderBy, boolean asc) {
+		if(ce == null) {// nothing have no children
+			return Collections.emptyList();
+		}
+		
 		StringBuilder sb = new StringBuilder();
 		sb.append("select cei from ").append(CatalogEntryImpl.class.getName()).append(" as cei ")
 		  .append(" inner join fetch cei.ownerGroup as ownerGroup")
+		  .append(" inner join fetch cei.parent as parentCei")
+		  .append(" inner join fetch parentCei.ownerGroup as parentOwnerGroup")
 		  .append(" left join fetch cei.repositoryEntry as repositoryEntry")
 		  .append(" left join fetch repositoryEntry.ownerGroup as repoOwnerGroup")
 		  .append(" left join fetch repositoryEntry.tutorGroup as repoTutorGroup")
 		  .append(" left join fetch repositoryEntry.participantGroup as repoParticipantGroup")
-		  .append(" where cei.parent.key=:parentKey");
-		if(orderBy != null) {
-			sb.append(" order by cei.").append(orderBy.name()).append(asc ? " ASC" : " DESC");
-		}
-
-		TypedQuery<CatalogEntry> dbQuery = dbInstance.getCurrentEntityManager()
-				.createQuery(sb.toString(), CatalogEntry.class)
-				.setParameter("parentKey", ce.getKey())
-				.setFirstResult(0);
-		if(maxResults > 0) {
-			dbQuery.setMaxResults(maxResults);
-		}
-
-		List<CatalogEntry> entries = dbQuery.getResultList();
-		return entries;
-	}
-	
-	public List<CatalogEntry> getShortChildrenOf(CatalogEntry ce, int firstResult, int maxResults, CatalogEntry.OrderBy orderBy, boolean asc) {
-		StringBuilder sb = new StringBuilder();
-		sb.append("select cei from ").append(CatalogEntryImpl.class.getName()).append(" as cei ")
-		  .append(" inner join fetch cei.ownerGroup as ownerGroup")
-		  .append(" left join fetch cei.repositoryEntry as repositoryEntry")
-		  .append(" left join fetch repositoryEntry.ownerGroup as repoOwnerGroup")
-		  .append(" left join fetch repositoryEntry.tutorGroup as repoTutorGroup")
-		  .append(" left join fetch repositoryEntry.participantGroup as repoParticipantGroup")
-		  .append(" where cei.parent.key=:parentKey");
+		  .append(" left join fetch repositoryEntry.olatResource as resource")
+		  .append(" where parentCei.key=:parentKey");
 		if(orderBy != null) {
 			sb.append(" order by cei.").append(orderBy.name()).append(asc ? " ASC" : " DESC");
 		}
@@ -178,15 +183,10 @@ public class CatalogManager extends BasicManager implements UserDataDeletable, I
 		StringBuilder sb = new StringBuilder();
 		sb.append("select cei from ").append(CatalogEntryImpl.class.getName()).append(" as cei ")
 		  .append(" inner join fetch cei.ownerGroup as ownerGroup")
-		  .append(" left join fetch cei.repositoryEntry as repositoryEntry")
-		  .append(" left join fetch repositoryEntry.ownerGroup as repoOwnerGroup")
-		  .append(" left join fetch repositoryEntry.tutorGroup as repoTutorGroup")
-		  .append(" left join fetch repositoryEntry.participantGroup as repoParticipantGroup")
-		  .append(" where cei.type= :type");
+		  .append(" where cei.type=").append(CatalogEntry.TYPE_NODE);
 		
 		return dbInstance.getCurrentEntityManager()
 				.createQuery(sb.toString(), CatalogEntry.class)
-				.setParameter("type", CatalogEntry.TYPE_NODE)
 				.getResultList();
 	}
 
@@ -292,16 +292,19 @@ public class CatalogManager extends BasicManager implements UserDataDeletable, I
 	 * @param ce
 	 */
 	public void deleteCatalogEntry(CatalogEntry ce) {
-		boolean debug = isLogDebugEnabled();
+		final boolean debug = isLogDebugEnabled();
 		if(debug) logDebug("deleteCatalogEntry start... ce=" + ce);
 		
 		if (ce.getType() == CatalogEntry.TYPE_LEAF) {
-			//delete catalog entry, then delete owner group
-			SecurityGroup owner = ce.getOwnerGroup();
-			dbInstance.getCurrentEntityManager().remove(ce);
-			if (owner != null) {
-				getLogger().debug("deleteCatalogEntry case_1: delete owner-group=" + owner);
-				securityManager.deleteSecurityGroup(owner);
+			//reload the detached catalog entry, delete it and then the owner group
+			ce = getCatalogEntryByKey(ce.getKey());
+			if(ce != null) {
+				SecurityGroup owner = ce.getOwnerGroup();
+				dbInstance.getCurrentEntityManager().remove(ce);
+				if (owner != null) {
+					getLogger().debug("deleteCatalogEntry case_1: delete owner-group=" + owner);
+					securityManager.deleteSecurityGroup(owner);
+				}
 			}
 		} else {
 			List<SecurityGroup> secGroupsToBeDeleted = new ArrayList<SecurityGroup>();
@@ -317,7 +320,7 @@ public class CatalogManager extends BasicManager implements UserDataDeletable, I
 			deleteCatalogSubtree(ce,secGroupsToBeDeleted);
 			// after deleting all entries, delete all secGroups corresponding
 			for (Iterator<SecurityGroup> iter = secGroupsToBeDeleted.iterator(); iter.hasNext();) {
-				SecurityGroup grp = (SecurityGroup) iter.next();
+				SecurityGroup grp = iter.next();
 				if(debug) logDebug("deleteCatalogEntry case_2: delete groups of deleteCatalogSubtree grp=" + grp);
 				securityManager.deleteSecurityGroup(grp);
 			}
@@ -368,15 +371,15 @@ public class CatalogManager extends BasicManager implements UserDataDeletable, I
 	 * @return List of catalog entries
 	 */
 	public List<CatalogEntry> getCatalogCategoriesFor(RepositoryEntry repoEntry) {
-		String sqlQuery = "select distinct parent from org.olat.catalog.CatalogEntryImpl as parent "
-			+ ", org.olat.catalog.CatalogEntryImpl as cei "
-			+ ", org.olat.repository.RepositoryEntry as re "
-			+ " where cei.repositoryEntry = re "
-			+ " and re.key= :reKey "
-			+ " and cei.parent = parent ";
-		
+		StringBuilder sb = new StringBuilder();
+		sb.append("select parentCei from ").append(CatalogEntryImpl.class.getName()).append(" as cei ")
+		  .append(" inner join cei.ownerGroup ownerGroup ")
+		  .append(" inner join cei.parent parentCei ")
+		  .append(" inner join fetch parentCei.ownerGroup parentOwnerGroup ")
+		  .append(" inner join cei.repositoryEntry re ")
+		  .append(" where re.key=:reKey");
 		return dbInstance.getCurrentEntityManager()
-				.createQuery(sqlQuery, CatalogEntry.class)
+				.createQuery(sb.toString(), CatalogEntry.class)
 				.setParameter("reKey", repoEntry.getKey())
 				.getResultList();
 	}
@@ -461,9 +464,12 @@ public class CatalogManager extends BasicManager implements UserDataDeletable, I
 	 * @return List of catalog entries
 	 */
 	public List<CatalogEntry> getRootCatalogEntries() {
-		String sqlQuery = "select cei from org.olat.catalog.CatalogEntryImpl as cei where cei.parent is null";
+		StringBuilder sb = new StringBuilder();
+		sb.append("select cei from ").append(CatalogEntryImpl.class.getName()).append(" as cei ")
+		  .append("inner join fetch cei.ownerGroup ownerGroup ")
+		  .append("where cei.parent is null");
 		return dbInstance.getCurrentEntityManager()
-				.createQuery(sqlQuery, CatalogEntry.class)
+				.createQuery(sb.toString(), CatalogEntry.class)
 				.getResultList();
 	}
